requirements.txt
# Core LLM and Document Processing
langchain==0.3.27
langchain-google-genai==2.0.4
google-generativeai==0.8.3
sentence-transformers==3.0.1

# Vector Database and Search  
faiss-cpu==1.8.0
chromadb==0.5.15

# Document Processing
pypdf==4.3.1
python-docx==1.1.2
pdfplumber==0.11.0
spacy==3.8.7

# Data Validation and Models
pydantic==2.9.0
pandas==2.3.1
numpy==1.26.4

# API and Interface
fastapi==0.115.0
streamlit==1.39.0
uvicorn==0.32.0
python-multipart==0.0.9

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
tqdm==4.67.0

# Testing
pytest==8.3.4
